Main project complete

+++ website/main/views.py
@@ -1,40 +1,45 @@
-import pandas as pd
-from django.contrib import messages
+import numpy as np
 from django.contrib.auth import login
-from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
-from django.http import HttpResponseRedirect, HttpResponse
-from django.shortcuts import redirect, render, get_object_or_404
+from django.http import HttpResponseRedirect
 from django.urls import reverse
-from django.views.decorators.http import require_POST
-from .forms import PrevisionForm, PrevisionInputForm, ChangeForm, NotificationForm
+from .forms import PrevisionInputForm, ChangeForm, NotificationForm
 from .forms import RegisterForm
 from .predictor import predict_scores
 from .forms import UserForm, ProfileForm
 from .models import Profile, Prevision, NewsletterSubscriber
-from .models import PredictionDataForm
 from .decorators import role_required
 from .models import Notification
 from django.contrib.staticfiles.storage import staticfiles_storage
-from django.shortcuts import render, get_object_or_404, redirect
+from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_POST
-from .models import PredictionDataForm
 from django.contrib import messages
+from .forms import PredictionDataForm
+from django.shortcuts import render, redirect
+from .forms import PrevisionForm
+import pandas as pd
+import numpy as np
+import json
 
 
 @login_required
 def list_data_entries(request):
-    data_entries = PredictionDataForm.objects.all()
+    data_entries = PredictionDataForm.objects.filter(validated=False)
     return render(request, 'main/validate_data.html', {'data_entries': data_entries})
+
 
 @login_required
 @require_POST
 def delete_data(request, entry_id):
     data_entry = get_object_or_404(PredictionDataForm, id=entry_id)
-    data_entry.delete()
-    messages.success(request, "Data deleted successfully.")
-    return redirect('list_data_entries')
+    if data_entry.validated == False:
+        data_entry.delete()
+        messages.success(request, "Data deleted successfully.")
+        return redirect('list_data_entries')
+    else:
+        return redirect('list_data_entries')
+
 
 @login_required
 @require_POST
@@ -44,7 +49,6 @@ def validate_data(request, entry_id):
     data_entry.save()
     messages.success(request, "Data validated successfully.")
     return redirect('list_data_entries')
-
 
 
 @login_required
@@ -58,6 +62,7 @@ def create_notification(request):
     else:
         form = NotificationForm()
     return render(request, 'main/create_notification.html', {'form': form})
+
 
 def notify_users(change):
     message = f"{change.user.username} updated the model '{change.name}' at {change.created_at.strftime('%Y-%m-%d %H:%M:%S')}: {change.description}"
@@ -77,10 +82,11 @@ def change_documentation(request):
             change.save()
             # Notify other users
             notify_users(change)
-            return redirect('home')  # Redirect to an appropriate page
+            return redirect('home')
     else:
         form = ChangeForm()
     return render(request, 'main/change_documentation.html', {'form': form})
+
 
 @login_required
 def view_notifications(request):
@@ -110,6 +116,7 @@ def data_input(request):
 
     return render(request, 'main/data_input.html', {'form': form})
 
+
 @login_required
 def edit_account(request):
     try:
@@ -124,7 +131,6 @@ def edit_account(request):
         if user_form.is_valid() and profile_form.is_valid():
             user_form.save()
             profile_form.save()
-            # Replace 'some_view' with the name of the view you want to redirect to after saving
             return redirect('profile')
     else:
         user_form = UserForm(instance=request.user)
@@ -136,14 +142,15 @@ def edit_account(request):
     })
 
 
-
 def base(request):
     lavender_image_url = staticfiles_storage.url('main/lavend.jpg')
     context = {'lavender_image_url': lavender_image_url}
     return render(request, 'main/base.html', context)
 
+
 def home(request):
     return render(request, 'main/home.html')
+
 
 def sign_up(request):
     if request.method == 'POST':
@@ -157,14 +164,18 @@ def sign_up(request):
 
     return render(request, 'registration/sign_up.html', {"form": form})
 
+
 def privacy_policy(request):
     return render(request, 'main/privacy_policy.html')
+
 
 def terms_of_service(request):
     return render(request, 'main/terms_of_service.html')
 
+
 def profile(request):
     return render(request, 'main/profile.html')
+
 
 @login_required
 def account_delete(request):
@@ -176,11 +187,14 @@ def account_delete(request):
 
     return redirect('main/profile.html')
 
+
 def convert_bool_string_to_numeric(df, boolean_fields):
+    pd.set_option('future.no_silent_downcasting', True)
+
     for field in boolean_fields:
-        # Convert fields in DataFrame from 'True'/'False' to 1/0
-        df[field] = df[field].replace({'True': 1, 'False': 0}).astype(int)
+        df[field] = df[field].replace({'True': True, 'False': False}).astype(bool).astype(int)
     return df
+
 
 def guest_prevision(request):
     if request.method == 'POST':
@@ -189,25 +203,20 @@ def guest_prevision(request):
             input_data = form.cleaned_data
             df_input = pd.DataFrame([input_data])
 
-            # Rename columns to match the expected format
             df_input.rename(columns={
-                'race_ethnicity': 'race/ethnicity', 
+                'race_ethnicity': 'race/ethnicity',
                 'parental_level_of_education': 'parental level of education'
             }, inplace=True)
 
-            # Specify the fields that should be treated as boolean
             boolean_fields = ['gender', 'lunch', 'test_preparation_course']
-            
-            # Convert boolean string values to numeric directly in df_input
+
             df_input = convert_bool_string_to_numeric(df_input, boolean_fields)
-            
-            # Proceed with one-hot encoding for categorical columns
+
             categorical_columns = ['race/ethnicity', 'parental level of education']
             df_input_encoded = pd.get_dummies(df_input, columns=categorical_columns)
-            
-            # Ensure the DataFrame matches the expected structure
+
             expected_columns = [
-                'gender', 'lunch', 'test preparation course','race/ethnicity_group A',
+                'gender', 'lunch', 'test preparation course', 'race/ethnicity_group A',
                 'race/ethnicity_group B', 'race/ethnicity_group C',
                 'race/ethnicity_group D', 'race/ethnicity_group E',
                 "parental level of education_associate's degree",
@@ -218,14 +227,12 @@ def guest_prevision(request):
                 'parental level of education_some high school'
             ]
 
-            df_input_encoded = df_input_encoded.reindex(columns=expected_columns, fill_value=0).astype('float32')
-            
-            # Use the prepared DataFrame for prediction
-            prediction_result = predict_scores(df_input_encoded)
-            # Assuming prediction_result is a dictionary with values that might be of type float32
-            prediction_result_serializable = {k: float(v) for k, v in prediction_result.items()}
+            df_input_encoded = df_input_encoded.reindex(columns=expected_columns, fill_value=0).astype('int32')
 
-            # Now, save the serializable version to the session
+            prediction_result = predict_scores(df_input_encoded)
+
+            prediction_result_serializable = {k: int(v) for k, v in prediction_result.items()}
+
             request.session['prediction_result'] = prediction_result_serializable
 
             return redirect('prediction_results')
@@ -238,19 +245,59 @@ def main_prevision(request):
     if request.method == 'POST':
         form = PrevisionForm(request.POST)
         if form.is_valid():
-            # Prepare data for prediction
-            prediction_data = {k: v for k, v in form.cleaned_data.items()}
-            request.session['prediction_data'] = prediction_data  # Store user input
+            input_data = form.cleaned_data
 
-            # Predict and store results
-            prediction_result = predict_scores(form.cleaned_data)  # Assumed function
-            request.session['prediction_result'] = prediction_result
+            # Create DataFrame from cleaned data
+            df_input = pd.DataFrame([input_data])
 
+            # Rename columns directly in the view function
+            df_input.rename(columns={
+                'race_ethnicity': 'race/ethnicity',
+                'parental_level_of_education': 'parental level of education'
+            }, inplace=True)
+
+            # Convert boolean fields from string representation to numeric
+            boolean_fields = ['gender', 'lunch', 'test_preparation_course']
+            for field in boolean_fields:
+                df_input[field] = df_input[field].replace({'True': 1, 'False': 0}).astype(int)
+
+            # One-hot encoding categorical columns
+            categorical_columns = ['race/ethnicity', 'parental level of education']
+            df_input_encoded = pd.get_dummies(df_input, columns=categorical_columns)
+
+            # Ensure DataFrame matches expected structure
+            expected_columns = [
+                'gender', 'lunch', 'test preparation course', 'race/ethnicity_group A',
+                'race/ethnicity_group B', 'race/ethnicity_group C',
+                'race/ethnicity_group D', 'race/ethnicity_group E',
+                "parental level of education_associate's degree",
+                "parental level of education_bachelor's degree",
+                'parental level of education_high school',
+                "parental level of education_master's degree",
+                'parental level of education_some college',
+                'parental level of education_some high school'
+            ]
+            df_input_encoded = df_input_encoded.reindex(columns=expected_columns, fill_value=0).astype('int32')
+
+            # Obtain prediction results
+            prediction_result = predict_scores(df_input_encoded)
+
+            # Ensure the prediction result is serializable
+            prediction_result_serializable = {k: int(v) for k, v in prediction_result.items()}
+
+            # Store serialized prediction result and input data in session
+            request.session['prediction_result'] = prediction_result_serializable
+            request.session['prediction_data'] = input_data
+
+            # Redirect to the prediction results page
             return redirect('prediction_results')
+        else:
+            # If the form is not valid, render the form with errors
+            return render(request, 'main/main_prevision.html', {'form': form})
     else:
+        # For GET requests, simply display the form
         form = PrevisionForm()
     return render(request, 'main/main_prevision.html', {'form': form})
-
 
 def prediction_results(request):
     prediction_result = request.session.get('prediction_result', {})
@@ -260,31 +307,31 @@ def prediction_results(request):
         'prediction_data': prediction_data
     })
 
-
-
-@login_required
 @login_required
 def saved_previsions(request):
     if request.method == 'POST':
-        prediction_data = request.session.get('prediction_data', {})
-        prediction_result = request.session.get('prediction_result', {})
+        if 'prediction_data' in request.session and 'prediction_result' in request.session:
+            prediction_data = request.session['prediction_data']
+            prediction_result = request.session['prediction_result']
 
-        new_prevision = Prevision.objects.create(
-            user=request.user,
-            gender=prediction_data['gender'],
-            lunch=prediction_data['lunch'],
-            test_preparation_course=prediction_data['test_preparation_course'],
-            race_ethnicity=prediction_data['race_ethnicity'],
-            parental_level_of_education=prediction_data['parental_level_of_education'],
-            math_score=prediction_result.get('math_score'),
-            reading_score=prediction_result.get('reading_score'),
-            writing_score=prediction_result.get('writing_score'),
-        )
-        return redirect('list_previsions')
+            new_prevision = Prevision.objects.create(
+                user=request.user,
+                gender=prediction_data['gender'],
+                lunch=prediction_data['lunch'],
+                test_preparation_course=prediction_data['test_preparation_course'],
+                race_ethnicity=prediction_data['race_ethnicity'],
+                parental_level_of_education=prediction_data['parental_level_of_education'],
+                math_score=prediction_result.get('math_score'),
+                reading_score=prediction_result.get('reading_score'),
+                writing_score=prediction_result.get('writing_score'),
+            )
+            print("New prediction saved:", new_prevision)
+            request.session.pop('prediction_data', None)
+            request.session.pop('prediction_result', None)
+            return redirect('main_prevision')
 
     previsions = Prevision.objects.filter(user=request.user)
     return render(request, 'main/saved_previsions.html', {'previsions': previsions})
-
 
 
 def subscribe_newsletter(request):
